requirements.txt
# Core dependencies
flask==2.3.3
flask-limiter==3.5.0
flask-sqlalchemy==3.1.1
flask-cors==4.0.0
sqlalchemy==2.0.23
redis==5.0.1

# Authentication
google-auth==2.25.2
pyjwt==2.8.0

# AI/ML
torch==2.1.2
transformers==4.36.2
sentence-transformers==2.3.1
openai==1.10.0
langchain==0.1.1

# Performance monitoring
psutil==5.9.6
gputil==1.4.0
py-spy==0.3.14

# System utilities
aiohttp==3.8.6
pyyaml==6.0.1
python-dotenv==1.0.0

# Plugin support
pluggy==1.3.0

# Optional NVIDIA GPU monitoring
# pynvml

# Optional dependencies for plugins
# googletrans==4.0.0-rc1
# requests-cache==1.1.0
# beautifulsoup4==4.12.2

# Development tools
pytest==7.4.3
black==23.11.0
isort==5.12.0
mypy==1.7.1
flake8==6.1.0
